Refactored the `getChecklistTaskUrls` selector
